+++ MCRA.Utils/Numerics/Matrix/LUDecomposition.cs
@@ -188,7 +188,7 @@ namespace MCRA.Utils {
             get {
                 double[] vals = new double[m];
                 for (int i = 0; i < m; i++) {
-                    vals[i] = (double)piv[i];
+                    vals[i] = piv[i];
                 }
                 return vals;
             }
@@ -208,7 +208,7 @@ namespace MCRA.Utils {
             if (m != n) {
                 throw new System.ArgumentException("Matrix must be square.");
             }
-            double d = (double)pivsign;
+            double d = pivsign;
             for (int j = 0; j < n; j++) {
                 d *= LU[j][j];
             }
